refactor(backup): extract boot swap into backup/swap.rs, move inline tests to tests/backup.rs
